Rename MapActuatorsStartupFilter, optimize middleware order (skip CF auth API call if not on management port)
