+++ src/Management/test/Endpoint.Test/ManagementHostBuilderExtensionsTest.cs
@@ -76,7 +76,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IDbMigrationsEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -102,7 +102,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IEnvironmentEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -127,7 +127,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         hostBuilder.AddHealthActuator();
         using IHost host = hostBuilder.Build();
 
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().HaveCount(1);
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().HaveCount(1);
         host.Services.GetService<IHealthAggregator>().Should().NotBeNull();
 
         await using AsyncServiceScope scope = host.Services.CreateAsyncScope();
@@ -149,7 +149,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         hostBuilder.ConfigureServices(services => services.AddHealthContributor<DownContributor>());
         using IHost host = hostBuilder.Build();
 
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().HaveCount(1);
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().HaveCount(1);
         host.Services.GetService<IHealthAggregator>().Should().NotBeNull();
 
         await using AsyncServiceScope scope = host.Services.CreateAsyncScope();
@@ -207,7 +207,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IHeapDumpEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -233,7 +233,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IActuatorEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -259,7 +259,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         Assert.NotNull(host.Services.GetService<IInfoEndpointHandler>());
-        Assert.NotNull(host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().FirstOrDefault());
+        Assert.NotNull(host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().FirstOrDefault());
     }
 
     [Fact]
@@ -271,7 +271,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         Assert.NotNull(host.Services.GetService<IInfoEndpointHandler>());
-        Assert.NotNull(host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().FirstOrDefault());
+        Assert.NotNull(host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().FirstOrDefault());
         Assert.NotNull(host.Services.GetServices<IInfoContributor>().OfType<TestInfoContributor>().FirstOrDefault());
     }
 
@@ -298,7 +298,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<ILoggersEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -352,7 +352,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<RouterMappings>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -378,7 +378,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IMetricsEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -404,7 +404,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IRefreshEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -430,7 +430,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IThreadDumpEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -456,7 +456,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IHttpExchangesEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -482,7 +482,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IServicesEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -508,7 +508,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<ICloudFoundryEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -555,7 +555,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using IHost host = hostBuilder.Build();
 
         host.Services.GetService<IActuatorEndpointHandler>().Should().NotBeNull();
-        host.Services.GetServices<IStartupFilter>().OfType<MapActuatorsStartupFilter>().Should().ContainSingle();
+        host.Services.GetServices<IStartupFilter>().OfType<ConfigureActuatorsMiddlewareStartupFilter>().Should().ContainSingle();
     }
 
     [Fact]
@@ -572,7 +572,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using HttpClient httpClient = host.GetTestClient();
 
         IStartupFilter[] startupFilters = host.Services.GetServices<IStartupFilter>().ToArray();
-        startupFilters.Should().ContainSingle(filter => filter is MapActuatorsStartupFilter);
+        startupFilters.Should().ContainSingle(filter => filter is ConfigureActuatorsMiddlewareStartupFilter);
         startupFilters.Should().ContainSingle(filter => filter is ManagementPortStartupFilter);
 
         HttpResponseMessage response = await httpClient.GetAsync(new Uri("/actuator", UriKind.Relative));
@@ -595,7 +595,7 @@ public sealed class ManagementHostBuilderExtensionsTest
             services.ConfigureActuatorEndpoints(endpoints => endpoints.RequireAuthorization("TestAuth"));
         });
 
-        // each of these will try to add their own MapActuatorsStartupFilter but should no-op in favor of the above
+        // each of these will try to add their own ConfigureActuatorsMiddlewareStartupFilter but should no-op in favor of the above
         hostBuilder.AddHypermediaActuator();
         hostBuilder.AddInfoActuator();
         hostBuilder.AddHealthActuator();
@@ -604,7 +604,7 @@ public sealed class ManagementHostBuilderExtensionsTest
         using HttpClient httpClient = host.GetTestClient();
 
         IStartupFilter[] startupFilters = host.Services.GetServices<IStartupFilter>().ToArray();
-        startupFilters.Should().ContainSingle(filter => filter is MapActuatorsStartupFilter);
+        startupFilters.Should().ContainSingle(filter => filter is ConfigureActuatorsMiddlewareStartupFilter);
         startupFilters.Should().ContainSingle(filter => filter is ManagementPortStartupFilter);
 
         // these requests hit the "RequireAuthorization" policy and will only pass if ConfigureWebHostWithSecureRouting is used
